make susi_linux hw independent (partial) (#477)
* main/states/led.py: mkae it work without spidev and /proc/asound

On desktops we will not have spidev, and in some testing scenarios
we wont have /proc/asound (chroot). Make led.py work under these
circumstances, too.

* install.sh: make youtube.lua installation arch independent

use dpkg-architecture -qDEB_HOST_MULTIARCH to determine the correct
lib triplet where the file should be installed.

* make state machine independent of GPIO module

move the loading of GPIO into the base state, and set a variable
if loading did not succeed. This way we don't try to reload again
and again in various states, but reuse the state variable.