fixes for overzealous detection of hotword
snowboy was listening all the time while playing back something,
and shot out a detection in case it was somehow found.

Stop the detector for short readings (say), and control turning on/off
the detector via the state change. That is, if we play music (which is
broken ATM) we transition back into the idle state and thus the
detector is running.